Atualização do GlobalState
