+++ pokedex/src/App.js
@@ -1,11 +1,14 @@
 import { Router } from './routes/Router';
 import './App.css';
 
+import GlobalState from "./global/GlobalState"
+
 function App() {
+
   return (
-    <div>
-    <Router />
-    </div>
+    <GlobalState>
+      <Router />
+    </GlobalState>
   );
 }
