+++ pokedex/src/global/GlobalContext.js
@@ -1,0 +1,3 @@
+import React, { createContext } from "react";
+
+export const GlobalContext = createContext();
