+++ pokedex/src/hooks/useRequestData.js
@@ -4,10 +4,7 @@ import axios from "axios";
 export const useRequestData = (url) => {
     const [data, setData] = useState(undefined)
 
-    useEffect(() => {
-        axios.get(url)
-        .then((res) => {setData(res.data.results)})
-        .catch((err) => {alert(err.message)})
-    }, [url])
+    
+    
     return data;
 }
